fix data bug

+++ app/src/main/java/com/fesvieira/whateverweather/models/WeatherData.kt
@@ -22,7 +22,7 @@ data class Location(
 data class CurrentWeather(
     val last_updated_epoch: Long,
     val last_updated: String,
-    val temp_c: Int,
+    val temp_c: Double,
     val temp_f: Double,
     val is_day: Int,
     val condition: Condition,
@@ -30,7 +30,7 @@ data class CurrentWeather(
     val wind_kph: Double,
     val wind_degree: Int,
     val wind_dir: String,
-    val pressure_mb: Int,
+    val pressure_mb: Double,
     val pressure_in: Double,
     val precip_mm: Double,
     val precip_in: Double,
